requirements_main.txt
accelerate==0.30.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.16.0
cloudpickle==3.1.2
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
datasets==2.19.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docker==7.1.0
emoji==2.12.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
fqdn==1.5.1
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
frozenlist==1.8.0
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.46
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
imbalanced-learn==0.12.3
importlib_metadata==7.2.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
langcodes==3.5.1
langdetect==1.0.9
lark==1.3.1
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.2.1
mistune==3.2.0
mlflow==2.13.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
murmurhash==1.0.15
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
onnxruntime==1.18.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optimum==1.19.2
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
preshed==3.0.13
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.2
pyarrow-hotfix==0.7
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2024.2
PyYAML==6.0.3
pyzmq==27.1.0
qtconsole==5.7.1
QtPy==2.4.3
querystring-parser==1.2.4
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
river==0.21.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.0
scipy==1.17.1
seaborn==0.13.2
Send2Trash==2.1.0
sentencepiece==0.2.1
setuptools==82.0.1
six==1.17.0
smart-open==6.4.0
smmap==5.0.3
soupsieve==2.8.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
sqlparse==0.5.5
srsly==2.5.3
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.19.1
torch==2.3.0+cpu
torchaudio==2.11.0+cpu
torchvision==0.18.0+cpu
tornado==6.5.5
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.0
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
vaderSentiment==3.3.2
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.3.4
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.7
wheel==0.46.3
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
